LayoutTests/inspector/canvas/create-context-webgpu-expected.txt

Test that CanvasManager tracks creation and destruction of WebGPU canvases.


== Running test suite: Canvas.CreateContextWebGPU
-- Running test case: Canvas.CreateContextWebGPU.NoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: Canvas.CreateContextWebGPU.Attached
PASS: Canvas context should be WebGPU.
  0: getContext - [native code]
  1: createAttachedCanvas - inspector/canvas/resources/create-context-utilities.js:4:36
  2: Global Code - [program code]
  3: evaluateWithScopeExtension - [native code]
  4: _evaluateOn - [native code]
  5: _evaluateAndWrap - [native code]

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGPU.Detached
PASS: Canvas context should be WebGPU.
  0: getContext - [native code]
  1: createDetachedCanvas - inspector/canvas/resources/create-context-utilities.js:8:62
  2: Global Code - [program code]
  3: evaluateWithScopeExtension - [native code]
  4: _evaluateOn - [native code]
  5: _evaluateAndWrap - [native code]

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGPU.CSSCanvas
Create CSS canvas from -webkit-canvas(css-canvas).
PASS: Canvas context should be WebGPU.
  0: getCSSCanvasContext - [native code]
  1: createCSSCanvas - inspector/canvas/resources/create-context-utilities.js:12:47
  2: Global Code - [program code]
  3: evaluateWithScopeExtension - [native code]
  4: _evaluateOn - [native code]
  5: _evaluateAndWrap - [native code]

PASS: Canvas name should equal the identifier passed to -webkit-canvas.
